Move styles to new component
